requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1.1
django-html5-boilerplate==1.0.8
django-shortcuts==1.6
idna==3.3
python-decouple==3.6
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
